Move NamedTensorLike to Torch.Typed.NamedTensor
