+++ hasktorch/test/Torch/Typed/NamedTensorSpec.hs
@@ -42,12 +42,13 @@ import qualified Torch.Tensor as D
 import Torch.Typed.Factories
 import Torch.Typed.Functional
 import Torch.Typed.Lens
+import Torch.Typed.NamedTensor
 import Torch.Typed.Tensor
 import Torch.Lens
 
-newtype Batch (n::Nat) a = Batch (Vector n a) deriving (Show, Eq, Generic)
-newtype Height (n::Nat) a = Height (Vector n a) deriving (Show, Eq, Generic)
-newtype Width (n::Nat) a = Width (Vector n a) deriving (Show, Eq, Generic)
+newtype Batch (n::Nat) a = Batch (Vector n a) deriving (Show, Eq)
+newtype Height (n::Nat) a = Height (Vector n a) deriving (Show, Eq)
+newtype Width (n::Nat) a = Width (Vector n a) deriving (Show, Eq)
 
 data RGB a = RGB
   { r :: a,
@@ -61,26 +62,10 @@ data YCoCg a = YCoCg
     co :: a,
     cg :: a
   }
-  deriving (Show, Eq, Generic)
+  deriving (Show, Eq, Generic, Default)
 
-data RGBA a = RGBA a a a a deriving (Show, Eq, Generic)
+data RGBA a = RGBA a a a a deriving (Show, Eq, Generic, Default)
 
-instance {-# OVERLAPS #-} (Coercible (vec n a) (Vector n a), KnownNat n, D.TensorLike (ToNestedList a), NamedTensorLike a) => NamedTensorLike (vec n a) where
-  type ToNestedList (vec n a) = [ToNestedList a]
-  toNestedList v = map (toNestedList @a) (V.toList (coerce v :: Vector n a))
-  asNamedTensor v = fromUnnamed . UnsafeMkTensor . D.asTensor $ toNestedList v
-  fromNestedList v = coerce (fmap fromNestedList . fromJust . V.fromList $ v :: Vector n a)
-  fromNamedTensor =  fromNestedList . D.asValue . toDynamic
-
-instance HasTypes Float Float where
-  types_ = id
-
-instance {-# OVERLAPS #-} (Generic (g a), Default (g a), HasTypes (g a) a, KnownNat (ToNat g), D.TensorLike (ToNestedList a), NamedTensorLike a) => NamedTensorLike (g a) where
-  type ToNestedList (g a) = [ToNestedList a]
-  toNestedList v = map (toNestedList @a) (flattenValues (types @a) v)
-  asNamedTensor v = fromUnnamed . UnsafeMkTensor . D.asTensor $ toNestedList v
-  fromNestedList v = replaceValues (types @a) def (fmap fromNestedList v)
-  fromNamedTensor =  fromNestedList . D.asValue . toDynamic
 
 testFieldLens :: HasField "r" shape => Lens' (NamedTensor '(D.CPU, 0) 'D.Float shape) (NamedTensor '(D.CPU, 0) 'D.Float (DropField "r" shape))
 testFieldLens = field @"r"
